requirements.txt
# TEC_OFFICE_REPO requirements

# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
typing>=3.7.4.3

# Gradio for HuggingFace Spaces interface
gradio>=3.50.0

# OpenAI API
openai>=1.0.0

# WordPress API
python-wordpress-xmlrpc>=2.3

# Storage and databases (expand as needed)
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional AI services
anthropic>=0.2.0

# For production deployment
gunicorn>=21.2.0
